requirements.txt
Django==2.0.1
django-bootstrap4==0.0.5
django-js-asset==1.0.0
django-mptt==0.9.0
django-taggit==0.22.2
psycopg2==2.7.3.2
pytz==2017.3
